@@ -1,25 +1,32 @@
 package ru.kata.spring.boot_security.demo.configs;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.authority.AuthorityUtils;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.Set;
+import java.util.List;
 
 @Component
 public class SuccessUserHandler implements AuthenticationSuccessHandler {
+    private static final Logger logger = LoggerFactory.getLogger(SuccessUserHandler.class);
+
     // Spring Security использует объект Authentication, пользователя авторизованной сессии.
     @Override
-    public void onAuthenticationSuccess(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse, Authentication authentication) throws IOException {
-        Set<String> roles = AuthorityUtils.authorityListToSet(authentication.getAuthorities());
-        if (roles.contains("ROLE_USER")) {
-            httpServletResponse.sendRedirect("/user");
+    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
+        List<String> roles = AuthorityUtils.authorityListToSet(authentication.getAuthorities()).stream().toList();
+        logger.info("User authenticated with roles: {}", roles);
+        if (roles.contains("ROLE_USER") && !request.getRequestURI().equals("/user")) {
+            logger.info("Redirecting to /user");
+            response.sendRedirect("/user");
         } else {
-            httpServletResponse.sendRedirect("/");
+            logger.info("Redirecting to /");
+            response.sendRedirect("/");
         }
     }
 }
